# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ERPGo API
  description: |
    A comprehensive Enterprise Resource Planning (ERP) system built with Go, providing user management, inventory management, order management, and business analytics capabilities.

    ## Overview
    ERPGo is a modern, cloud-native ERP system designed for high performance, scalability, and maintainability. It provides:

    - User authentication and authorization with JWT tokens
    - Product catalog management with categories and variants
    - Inventory tracking and warehouse management
    - Order processing and fulfillment workflows
    - Customer management and address handling
    - Comprehensive analytics and reporting

    ## Base URLs
    - **Production**: `https://api.erpgo.example.com`
    - **Staging**: `https://api-staging.erpgo.example.com`
    - **Development**: `http://localhost:8080`

    ## Authentication
    The API uses JWT (JSON Web Tokens) for authentication. Include the token in the `Authorization` header:

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting
    API requests are rate-limited to ensure fair usage:
    - **Default**: 100 requests per minute
    - **Burst**: 200 requests
    - **Window**: 1 minute

    Rate limit headers are included in responses:
    ```
    X-RateLimit-Limit: 100
    X-RateLimit-Remaining: 99
    X-RateLimit-Reset: 1640995200
    ```
  version: 1.0.0
  contact:
    name: ERPGo Support
    email: support@erpgo.example.com
...
  license:
...
servers:
...
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "user@example.com"
              password: "password123"
              remember: true
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          example: "password123"
        remember:
          type: boolean
          default: false
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expires_at:
          type: string
          format: date-time
          example: "2024-01-09T12:00:00Z"
        user:
          $ref: '#/components/schemas/UserInfo'
...
    UserInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "user@example.com"
        username:
          type: string
          example: "username"
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        phone:
          type: string
          example: "+1234567890"
        roles:
          type: array
          items:
            type: string
          example: ["user", "admin"]
        is_active:
          type: boolean
          example: true
        is_verified:
          type: boolean
          example: true
        last_login_at:
          type: string
          format: date-time
          example: "2024-01-08T12:00:00Z"
        created_at:
          type: string
          format: date-time
          example: "2024-01-08T10:30:00Z"
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Request validation failed"
            details:
              oneOf:
                - type: array
                  items:
                    type: object
                    properties:
                      field:
                        type: string
                        example: "email"
                      message:
                        type: string
                        example: "Invalid email format"
                - type: string
                  example: "Additional error details"
            timestamp:
              type: string
              format: date-time
              example: "2024-01-08T10:30:00Z"
            request_id:
              type: string
              example: "req_123456789"
  responses:
    BadRequest:
      description: Bad request - Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "BAD_REQUEST"
              message: "Invalid request data"
              timestamp: "2024-01-08T10:30:00Z"
              request_id: "req_123456789"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
              timestamp: "2024-01-08T10:30:00Z"
              request_id: "req_123456789"
...
    ValidationError:
      description: Validation error - Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Request validation failed"
              details:
                - field: "email"
                  message: "Invalid email format"
              timestamp: "2024-01-08T10:30:00Z"
              request_id: "req_123456789"
...